requirements.txt
discord.py
requests
PIL